Ignore the dist folder when testing on Codeship
Jest has declared that it will no longer allow multiple package.json
files with the same package name to exist in a repo. I'm still not sure
why this is only failing now. Since codeship runs our build before the
test it means there is a dist/ folder sitting there with a copy of our
server in it. Jest is getting mad that we're trying to test two packages
with the same name. This ignore _should_ mean that it wont auto discover
test files in the dist directory.